issues discovered from modding `timestep/src/timer.js`
Ran a translation on timestep to see how things were coming, and found a couple of issues.  This was a pretty bad one (noticed it by accident).

Example mod output:
``` js
// ...
export const now = 0;
export const frames = 0;

export const tick = function (dt) {
  // ...
  export const now = dt;
  exports.frames++;
  exports.onTick(dt);
  // ...
};
// ...
```

One of the problems is with how babel transforms the es6 exports, see an example here:
<http://babeljs.io/repl/#?evaluate=true&lineWrap=true&presets=es2015%2Ces2015-loose%2Creact%2Cstage-0%2Cstage-1%2Cstage-2%2Cstage-3&code=export%20let%20now%20%3D%200%3B%0A%0Aexport%20const%20tick%20%3D%20function%20(dt)%20%7B%0A%20%20%2F%2F%20Translates%20correctly%20(both%20module%20level%20and%20exports%20%60now%60%20are%20updated)%0A%20%20now%20%2B%3D%20dt%3B%0A%20%20%2F%2F%20Translates%20incorrectly%20(module%20level%20%60now%60%20is%20not%20updated)%0A%20%20exports.now%20%2B%3D%20dt%3B%0A%7D%3B%0A%0Aconsole.log(now)%0A>

There are another couple of minor things exposed here:

- `export` appearing in a function block
- `export const` of `now` more than once
- `export const` of variable that should be `let`
- `export const now` instead of writing to the `now` or `exports.now` variables
- `not += dt` incorrectly translating to `now = dt`

My proposal is that if there is a naming collision, we warn.  The code is technically still valid and will run, but does not look as nice as we want.

`writeToExportedVar`: exposes a few issues with the current export transform.
`writeToExportedVarCollision`: shows a similar scenario where we should warn the user, and not transform code.